- Utxo set    * model, services, DB    * sync up to 82K - APIs    * /address/summary

+++ src/main/java/com/btc/api/dao/GenericDAO.java
@@ -43,7 +43,7 @@ public class GenericDAO {
     }
 
     @Transactional
-    public <E> int count(Class<E> clazz) {
+    public <E> long count(Class<E> clazz) {
         return list(clazz).size();
     }
